vfio: fix brokenness
Trying to re-use the same file for all device/iommu fds is wrong, just
use anon_inode_getfd().  This blew up when we had two devices opened and
killed the test tool.

We can close the iommu of there are no open devices attached to it and
the iommu isn't open, but don't look at the group refcnt.  This prevented
a different process from opening a device after the original owner
closed it, but still had the group open.

Do look at the group refcnt before unsetting the iommu and freeing it.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>