vfio: Fixup private API
Now that PCI is included, it's obvious that letting the device driver
allocate the vfio_device is a waste of space.  We end up with vfio
driver private data sitting around for devices that aren't bound to
their vfio device driver.  Let's get rid of the alloc/free device ops
interface and let the driver allocate it's own data on bind and free
on remove.  We'll pass this as opaque data into vfio, which vfio
will return to the driver on file operations.  Then we can move
vfio API defines out of vfio_private.h and into vfio.h, and vfio
device drivers have no need to get to vfio_device or vfio_iommu.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>