vfio: IOMMU cleanups
I had skipped the unmapall function, thinking we don't need it.
When we free the iommu domain, it tears down all the mappings, but
we still need to walk the vfio_iommu structures and free them.
This could probably be optimized to just free instead of unmap
and free, but re-use the old code for now.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>