vfio: Add group and iommu flags
It's useful for userspace to have an easy way to tell whether a
group is viable and/or locked so they can figure out why their
GET_*_FD commands aren't working.  For IOMMU, we currently assume
we can map anything.  Make this explicit so we can later add a
different flag for other types of IOMMUs.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>