pci: Translate pcibios errors to standard errno
Several pci functions return PCIBIOS_* values (0/>0).  Not only are
we inconsistent in using this versus errno across architectures,
sometimes we want to return an error through non-PCI #defines.  Add
a simple inline function to replace the PCI specific error with the
most appropriate errno.  This is safe to call whether the error
value is already an errno or PCIBIOS_*.

Xen already has a function of the same name, but translates to a
Xen specific error value scheme.  Trivial rename for them to
avoid conflict.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>