vfio: update to new iommu groups
This is a first pass hack at updating vfio & vfio-pci to use the
new iommu group interface.  Much cleanup necessary.  The significant
change is that since vfio is no longer managing groups, we don't need
to expose all of the groups in the system.  Therefore rather than
vfio-pci registering every device with vfio at init, we now only
register vfio groups if one of the devices in the group is attached
to a vfio driver.  For example, initially /dev/vfio doesn't exist.
If device 0000:01:10.1 belongs to group 42 and I bind this device to
vfio-pci, /dev/vfio/42 appears.  We could enumerate every group, but
doing it on demand feels like the right way.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>