vfio: rename main file, change iommu driver ops
Since the vfio core infrastructure is now a single file, drop the
_main and just call it vfio.c.

Changed attach/detach group to be a separate callback on iommu
driver ops.  We cannot allow userspace access to this or else the
user would have free access to extract their group from the iommu.
we therefore pull it out to a separate function so that we don't
need to filter ioctls.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>