pci: Add Ricoh quirk and generic ACS support for iommu groups
It's long been requested that iommu groups enable iommu drivers
to support dma quirks for devices.  One such quirk is a Ricoh
multifunction device where dma from all the function appears to
be from function 0.  iommu drivers therefore need to group all
functions together.  This is easily done by lookup up the dma
quirk for a specific device.  Making use of this when creating
iommu domains is left for the iommu drivers.

PCI ACS is another limiter for isolation.  If ACS is not enabled
all the way up the chain from a device to the iommu then dma from
the device can be redirected without reaching the iommu.  This
therefore presents visibility issues to the iommu.  Much like the
dma quirk, introduce a callback that returns the highest level
device with ACS support enabled.  iommu groups can then take this
into account such that the group cannot dma to memory or other
devices without passing through the iommu.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>