pci: Add ACS validation utility
In a PCIe environment, transactions aren't always required to
reach the root bus before being re-routed.  Intermediate
switches between an endpoint and the root bus can redirect
DMA back downstream before things like IOMMU have a chance
to intervene.  This utility function allows us to determine
the closest device for which ACS is enabled back to the root
bus for use in determining the boundaries of an iommu group.
Logic for this extracted from libvirt.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>