kvm: Extend irqfd to support level interrupts
KVM_IRQFD currently only supports edge triggered interrupts,
asserting then immediately deasserting an interrupt.  There are a
couple ways we can emulate level triggered interrupts using
discrete events depending on the usage model we expect from drivers.
This patch implements a level emulation model useful for external
assigned device drivers, like VFIO.  The irqfd is used to assert
the interrupt.  When the guest issues an EOI for the interrupt, the
level is automatically deasserted and the irqfd user is notified via
an eventfd.  This is therefore the LEVEL_EOI extension to KVM_IRQFD.
To do this, we need to allocate a new irq source ID for the interrupt
so we don't get interference from userspace.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>