vfio-pci: Add PCI bus reset ioctl
We have a variety of reset mechanisms for performing a function level
reset, but devices need to opt-in to supporting these to make them
useful.  Graphics devices in particular have poor support for FLR and
don't seem to do much of any reset on D3hot->D0 transition.  This
means graphics devices are often a one-shot (or less) endeavour for
userspace, after which point the card is in an unknown state and hard
for the next driver to recover.

Discrete graphics devices are almost always below a PCIe root port
(PCI-to-PCI bridge), which gives us the opportunity to remove the
device dependency and reset via a secondary bus reset. A complication
here is that a but may have multiple devices, each potentially in
separate IOMMU groups.  Obviously we can't reset a bus (and all
subordinate buses) unless the user is in control of all of those
devices.  We therefore make the requirement that in order to perform
a bus reset, all of the groups affected by the reset must be attached
to the same container.  Therefor if 1:00.0 and 1:00.1 are functions
within the same IOMMU group and device 1:00.0 is being accessed by
userspace, a reset of bus 1 is available through device 1:00.0
because 1:00.1 is implicitly owned by the user.  If 1:00.0 and 1:00.1
are in separate groups (ACS supported), we can no longer make that
assumption and the user must explicity add both groups to the same
container to make a bus reset available.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>