PCI: quirk pci_for_each_dma_alias()
Some devices are broken and use a requester ID other than their
physical devfn.  Add a byte, using an existing gap in the pci_dev
structure, to store an alternate "alias" devfn.  A bit in the
dev_flags tells us when this is valid.  We then add the alias as
one more step in the pci_for_each_dma_alias() iterator.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>