PCI: quirk dma_alias_devfn for Ricoh devices
The existing quirk for these devices doesn't really solve the problem;
re-implement it using the DMA alias iterator.  We'll come back later
and remove the existing quirk and dma_source interface.  Note that
device ID 0xe822 is typically function 0 and 0xe230 has been tested to
not need the quirk and are therefore removed versus the equivalent
dma_source quirk.  If there exist in other configurations we can
re-add them.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>