PCI: Extend and export pci_msi_supported()
Support for multiple MSI vectors is pretty much limited to x86
systems with interrupt remapping hardware.  Unfortunately drivers
have no way to predict whether such hardware is there until they try
to request multiple vectors and get a failure.  For drivers like
vfio-pci, we can be smarter about what we expose if we know the
platform can only support a single vector.  For example, we can
virtualize the MSI capability exposed to userspace to reflect this
limitation.  In a virtual machine environment, the guest driver is
far more likely to work with a single vector if that's all we
advertise as available.

This extends and exports the existing pci_msi_supported() function
to provide this feature and backs the x86 interrupt remapping with
the necessary handler.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>