vfio/pci: Fixup MSI multiple message capability
I'm not sure what we were doing with msi_qmax previously, some sort
of non-spec defined updating, but it doesn't work well with some
devices.  An Atheros 93xx supports 4 MSI vectors and the Windows
driver attempts to enable them all.  We can only support more than
a single vector on x86 platforms with interrupt remapping.  If we
don't have that, the 4 vector allocation fails and the guest gets 1.
Sometimes that works, even with this Atheros device, but often it
doesn't.  If we know we can only support a single MSI vector, let's
just advertise that as capability.  This seems to let the Atheros
device work consistently and traces show that the driver is enabling
a single vector.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>