vfio-pci: Allow PCI IDs to be specified as module options
This copies the same support from pci-stub for exactly the same
purpose, enabling a set of PCI IDs to be automatically added to the
driver's dynamic ID table.  The code here is pretty simple and both
vfio-pci and pci-stub are fairly unique in being meta drivers, capable
of attaching to any device, so there's no attempt made to generalize
the code into pci-core.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>