nouveau/radeon: Un-ignore hotplug on resume
Commit b440bde74f04 enabled drivers to be able to set an ignore
hotplug flag on the device and used it in the suspend routine for
nouveau and radeon, but no way was provided to clear the flag,
allowing it to remain as driver cruft on the device.  Now that we have
a function to clear the flag, use it in the resume routine of these
same drivers.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>